Seperate authentication model in different app
