+++ techeria_controller/.env
@@ -1,0 +1,2 @@
+export EMAIL_HOST_USER = kabirulislam.civil@gmasil.com
+export EMAIL_HOST_PASSWORD = Jaijaidin
