+++ techeria_controller/techeria_app/models.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.models import AbstractUser
 from django.db import models
 from django.db.models.fields import EmailField
 
@@ -15,40 +16,7 @@ class Products(models.Model):
         db_table = 'products'
 
 
-class BuyerModel(models.Model):
-    first_name = models.CharField(max_length=255)
-    last_name = models.CharField(max_length=255)
-    user_name = models.CharField(max_length=255)
-    date_of_birth = models.DateField()
-    email = models.EmailField()
-    mobile_number = models.IntegerField(max_length=10)
-    address = models.CharField(max_length=255)
-    city = models.CharField(max_length=255)
-    state = models.CharField(max_length=255)
-    zip_code = models.CharField(max_length=255)
-    country = models.CharField(max_length=20)
 
-
-    class Meta:
-        db_table = "buyer"
-
-
-class SellerModel(models.Model):
-    first_name = models.CharField(max_length=255)
-    last_name = models.CharField(max_length=255)
-    user_name = models.CharField(max_length=255)
-    date_of_birth = models.DateField()
-    email = models.EmailField()
-    mobile_number = models.IntegerField(max_length=10)
-    address = models.CharField(max_length=255)
-    city = models.CharField(max_length=255)
-    state = models.CharField(max_length=255)
-    zip_code = models.CharField(max_length=255)
-    country = models.CharField(max_length=20)
-
-
-    class Meta:
-        db_table = "seller"
 
 
 class Laptops(models.Model):
@@ -93,5 +61,5 @@ class Smartphone(models.Model):
 
 
 
-    def __str__(self):
-        return self.techeria_app
+def __str__(self):
+    return self.techeria_app
