+++ techeria_controller/techeria_controller/settings.py
@@ -1,3 +1,5 @@
+
+   
 """
 Django settings for techeria_controller project.
 Generated by 'django-admin startproject' using Django 3.2.8.
