+++ techeria_controller/techeria_controller/urls.py
@@ -15,24 +15,16 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
+from django.urls.conf import include
 from techeria_app import views
+
 
 from django.conf import settings
 from django.conf.urls.static import static
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', views.index, name='index'),
-    path('about', views.about, name='about'),
-    path('contact', views.contact, name='contact'),
-    path('laptop', views.laptop, name='laptop'),
-    path('smartphone', views.smartphone, name='smartphone'),
-    path('loginpage', views.loginpage, name='loginpage'),
-    path('registration', views.registration, name='registration'),
-    path('product', views.product, name='product'),
-    path('cart', views.cart, name='cart'),
-    path('checkout/', views.checkout, name='checkout'),
-    path('search', views.search, name='search'),
-    path('p', views.productInfo, name='p'),
-     path('logout', views.logout, name='logout'),
-] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+    path('', include('techeria_app.urls')),
+    path('', include('authentication.urls')),
+
+]
