+++ techeria_controller/templates/registration.html
@@ -85,7 +85,7 @@ body {
 <tr>
 <td>MOBILE NUMBER</td>
 <td>
-<input type="Number" name="Mobile_Number" />
+<input type="Number" name="Mobile_Number" min="-9999999999" max="9999999999" />
 (10 digit number)
 </td>
 </tr>
